backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
supabase>=2.3.0
firebase-admin>=6.2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.2
